# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: 'LawLens Morocco API Documentation'
  description: 'API de LawLens Morocco — accompagnement juridique des entrepreneurs au Maroc (Sanctum).'
  version: 1.0.0
servers:
  -
    url: 'http://localhost:8000'
...
components:
  securitySchemes:
    default:
      type: http
      scheme: bearer
      description: "Récupérez un jeton en vous authentifiant via <b>POST /api/login</b> (ou <b>POST /api/register</b>), puis envoyez-le dans l'en-tête <code>Authorization: Bearer &lt;token&gt;</code>."
security:
  -
    default: []
paths:
...
  /api/login:
    post:
      summary: Connexion
      operationId: connexion
      description: 'Retourne le profil utilisateur et un jeton Sanctum.'
      parameters: []
      responses:
        200:
          description: success
          content:
            application/json:
              schema:
                type: object
                example:
                  user:
                    id: 1
                    name: 'Yasmine Alaoui'
                    email: yasmine@example.com
                    role: entrepreneur
                  token: 1|abc123...
                properties:
                  user:
                    type: object
                    properties:
                      id:
                        type: integer
                        example: 1
                      name:
                        type: string
                        example: 'Yasmine Alaoui'
                      email:
                        type: string
                        example: yasmine@example.com
                      role:
                        type: string
                        example: entrepreneur
                  token:
                    type: string
                    example: 1|abc123...
        401:
          description: 'invalid credentials'
          content:
            application/json:
              schema:
                type: object
                example:
                  message: 'Identifiants invalides.'
                properties:
                  message:
                    type: string
                    example: 'Identifiants invalides.'
      tags:
        - Endpoints
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  description: validation.email.
                  example: gbailey@example.net
                password:
                  type: string
                  description: ''
                  example: '|]|{+-'
              required:
                - email
                - password
      security: []
...